Refactor to use environment variables for configuration
Remove arguments and use environment variables to configure the application instead. This simplifies the operation and removes the complicated argument parsing logic.

This also removes using anyhow and uses a custom error enum instead. Anyhow is great but not needed here, cause the app is so simple. This shaves another 200k off the release biniary size.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>