Handle response status codes correctly
Response status codes were not being handled as expected. Previously, ureq would handle 4xx & 5xx, so those worked but this was accidental. It would not make 3xx an error either and the logic to do that was flawed, so those would succeed.

This PR makes a change so that we handle status codes as follows:

- 2xx is OK
- 3xx is FAIL
- 4xx & 5xx FAIL (and are properly converted to the expected error

Signed-off-by: Daniel Mikusa <dan@mikusa.com>